+++ deps/icu-small/source/tools/toolutil/pkg_genc.cpp
@@ -418,9 +418,9 @@ writeCCode(
         filename,
         destdir,
         buffer,
-        sizeof(buffer),
+        static_cast<int32_t>(sizeof(buffer)),
         entry + uprv_strlen(entry),
-        sizeof(entry) - uprv_strlen(entry),
+        static_cast<int32_t>(sizeof(entry) - uprv_strlen(entry)),
         ".c",
         optFilename);
 
@@ -679,7 +679,7 @@ getOutFilename(
         outFilenameBuilder.append(destdir, status);
         outFilenameBuilder.ensureEndsWithFileSeparator(status);
     } else {
-        outFilenameBuilder.append(inFilename, basename - inFilename, status);
+        outFilenameBuilder.append(inFilename, static_cast<int32_t>(basename - inFilename), status);
     }
     inFilename=basename;
